refactor(table_functions): move to functions
