+++ query/src/functions/table_functions/table_function.rs
