+++ query/src/storages/memory/numbers_table.rs
@@ -35,6 +35,8 @@ use common_planners::Statistics;
 use common_streams::SendableDataBlockStream;
 
 use super::numbers_table_stream::NumbersStream;
+use crate::functions::TableArgs;
+use crate::functions::TableFunction;
 use crate::pipelines::new::processors::port::OutputPort;
 use crate::pipelines::new::processors::processor::ProcessorPtr;
 use crate::pipelines::new::processors::SyncSource;
@@ -44,8 +46,6 @@ use crate::pipelines::new::SourcePipeBuilder;
 use crate::pipelines::transforms::get_sort_descriptions;
 use crate::sessions::QueryContext;
 use crate::storages::Table;
-use crate::table_functions::TableArgs;
-use crate::table_functions::TableFunction;
 
 pub struct NumbersTable {
     table_info: TableInfo,
