+++ query/tests/it/storages/fuse/table_test_fixture.rs
@@ -27,6 +27,7 @@ use common_planners::Expression;
 use common_planners::Extras;
 use common_streams::SendableDataBlockStream;
 use databend_query::catalogs::Catalog;
+use databend_query::functions::TableArgs;
 use databend_query::interpreters::CreateTableInterpreter;
 use databend_query::interpreters::InterpreterFactory;
 use databend_query::sessions::QueryContext;
@@ -39,7 +40,6 @@ use databend_query::storages::fuse::FUSE_TBL_SNAPSHOT_PREFIX;
 use databend_query::storages::Table;
 use databend_query::storages::ToReadDataSourcePlan;
 use databend_query::storages::OPT_KEY_DATABASE_ID;
-use databend_query::table_functions::TableArgs;
 use futures::TryStreamExt;
 use tempfile::TempDir;
 use uuid::Uuid;
